Render placeholder text as HTML tags

Renders ipsum() placeholder text as HTML tags using htmltools::as.tags(). By default, paragraphs are rendering <p> tags with htmltools::p(), but you may provide your own paragraph wrapper function. In general, you won't need to manually call this function, instead you can just drop ipsum() text into another htmltools::tag().

Usage

# S3 method for lorem
as.tags(x, wrapper = NULL, ...)

Arguments

x

Object to be converted.

wrapper

A function that takes a character string of a paragraph or chunk of placeholder text and returns an htmltools::tag(). By default, lorem ipsum text is wrapped in htmltools::p().

...

Any additional parameters.

Value

Returns an htmltools::tagList().

Examples

htmltools::div(
  lorem::ipsum(paragraphs = 3)
)
#> <div>
#>   <p>Dolor justo sagittis tristique libero non sed fames maecenas. Viverra sollicitudin eu odio ullamcorper sem sociis porta volutpat facilisi. Fermentum dignissim sodales diam sapien fusce sagittis aenean.</p>
#>   <p>Dolor dignissim ligula dapibus vel faucibus nibh bibendum. Phasellus elementum molestie elementum diam, turpis curabitur diam proin justo purus. Suscipit laoreet, inceptos duis: phasellus donec iaculis in sodales. Justo posuere parturient litora nullam nec pharetra aliquam vitae vel proin!</p>
#>   <p>Elit purus suscipit odio viverra, purus nostra sodales torquent nisl mus. Cursus magna netus praesent velit; leo massa, per, posuere sed nulla etiam bibendum. Sollicitudin justo lacinia mollis mauris blandit platea, ante habitasse egestas consequat. Duis pretium aliquam pulvinar tristique. Nunc netus sapien nec vulputate metus.</p>
#> </div>

htmltools::tags$ul(
  htmltools::as.tags(
    lorem::ipsum(paragraphs = 3, sentences = 1),
    wrapper = htmltools::tags$li
  )
)
#> <ul>
#>   <li>Elit metus tempor quisque elementum aenean ligula gravida bibendum nibh inceptos nisi nunc?</li>
#>   <li>Consectetur justo vehicula non eleifend magna curae fusce volutpat sed.</li>
#>   <li>Lorem nostra feugiat sem nisl parturient at donec himenaeos habitasse facilisi cubilia gravida!</li>
#> </ul>