Generate Lorem Ipsum Text

Generates lorem ipsum placeholder text for the requested number of sentences or paragraphs. You can control the number of sentences per paragraph and the average number of words per sentence, or simply enter the number of desired paragraphs for a completely random experience.

lorem::ipsum() uses sampling and the random number generator and makes no effort to shield the placeholder text generation from the main script, so please only use this package for temporary placeholder text.

Usage

ipsum(paragraphs = 1, sentences = NULL, avg_words_per_sentence = 10)

ipsum_words(n, collapse = TRUE)

ipsum_starts(n)

Arguments

paragraphs

Number of paragraphs of text to generate.

sentences

Number of sentences per paragraph. If NULL, then a random number of sentences per paragraph (approximately 3-8) will be chosen. Alternatively, sentences can be a vector of integers representing the number of sentences per paragraph.

avg_words_per_sentence

Number of expected words per sentence.

n

Number of words to generate

collapse

Should the words be collapsed into a single string, separated by spaces (default)? If FALSE, the chosen words are returned as a character vector.

Value

A character vector of lorem ispum placeholder text, where each element in the vector is a paragraph of text.

Functions

ipsum(): Generate paragraphs and sentences of lorem ipsum text.

ipsum_words(): Generate lorem ipsum words, without punctuation.

ipsum_starts(): Generate lorem ipsum starting words.

Options

You can influence, to a degree, the amount of punctuation that is included in the output using the lorem.punctuation_valence option. This global option should be a number between 0 and 1, or FALSE to disable punctuation altogether. When the value is closer to 1, more punctuation is included in the sentences. When the value is closer to 0, less punctuation will be inserted. The default value is 0.4.

Examples

# 1 paragraph of text
lorem::ipsum(1)
#> <lorem>
#> Elit luctus interdum cras condimentum? Orci etiam in taciti venenatis
#> dis dui luctus viverra dapibus! Litora taciti vulputate volutpat:
#> aptent tempor est ante pellentesque. Augue porttitor etiam litora
#> cursus facilisi: turpis facilisi praesent, libero augue eleifend ante.
#> Pharetra turpis cursus orci, laoreet conubia feugiat ridiculus nisl ac:
#> tortor quam venenatis at dictum enim torquent pretium nunc!

# 2 paragraphs with 2 and 3 sentences each
lorem::ipsum(2, sentences = c(2, 3))
#> <lorem>
#> Amet molestie habitasse a habitant phasellus quam consequat. Potenti
#> fames faucibus condimentum vel mollis dictum posuere nostra semper nam
#> nascetur gravida.
#> 
#> Lorem nostra eleifend feugiat duis dictumst litora dictum etiam.
#> Sodales suscipit mus proin arcu suscipit imperdiet! Imperdiet
#> sollicitudin felis risus nostra imperdiet neque hendrerit nascetur
#> accumsan massa at nec velit fringilla sollicitudin sollicitudin montes
#> vel massa?

# 2 paragraphs with short sentences
lorem::ipsum(2, avg_words_per_sentence = 4)
#> <lorem>
#> Sit platea taciti est ac. Parturient purus, curabitur class etiam.
#> Natoque pharetra proin suspendisse sed. Inceptos gravida aenean nisl
#> nisi. Morbi ultricies ridiculus sapien euismod nunc! Erat tellus fusce;
#> dictum posuere ultricies. Na ultricies.
#> 
#> Sit nostra praesent porta? Auctor porta ultrices hac. Viverra sapien
#> cubilia id. Pharetra lacinia sollicitudin vestibulum. Na vestibulum.

# 2 paragraphs with long sentences
lorem::ipsum(2, avg_words_per_sentence = 20)
#> <lorem>
#> Adipiscing sollicitudin habitant placerat egestas cum. Sodales netus
#> duis potenti nisl, tristique volutpat duis vestibulum tortor odio?
#> Commodo taciti bibendum taciti, lectus primis luctus lacinia convallis.
#> Ullamcorper bibendum vulputate, ante class etiam volutpat nisl duis?
#> Sem id class senectus donec, parturient quisque vestibulum tortor
#> aptent litora. Duis id purus fusce fames eleifend bibendum aliquam
#> integer neque bibendum augue sapien eros sollicitudin fermentum
#> placerat natoque ac nisl eget varius egestas donec commodo dictumst
#> himenaeos natoque urna mollis etiam, tortor tristique primis faucibus
#> interdum vitae, senectus a felis, lectus curabitur, mollis mus netus
#> class, mattis, tempus – elementum, purus, purus aliquet sociosqu
#> hendrerit eu integer facilisi proin dui diam libero eu elementum nulla
#> cursus ultricies placerat luctus porttitor curabitur aenean himenaeos
#> metus quis ultricies mus odio?
#> 
#> Dolor non inceptos egestas potenti varius lobortis nibh tortor urna!
#> Donec senectus quam sociosqu cursus eget auctor eget proin. Purus sed –
#> phasellus; vestibulum accumsan mus natoque, tempor conubia placerat?
#> Pulvinar consequat laoreet etiam tincidunt. Conubia lacinia lectus
#> phasellus mus ultrices arcu blandit. Mi est facilisis faucibus vitae
#> mus urna parturient pretium. Urna auctor sed rhoncus dictumst sodales
#> neque facilisi ultricies tristique euismod nisi diam fames egestas
#> nascetur arcu non cursus ullamcorper, dapibus curae ultricies vehicula
#> mauris quisque dictum: sem sollicitudin, posuere non cursus cubilia
#> etiam felis cursus, eu, suscipit gravida magna fames vel hac
#> pellentesque, elementum: rhoncus bibendum lacinia, lectus id mattis
#> vestibulum proin posuere, etiam, nascetur diam suscipit, interdum
#> faucibus id odio molestie bibendum fames eros, per hendrerit felis
#> vestibulum phasellus sociosqu aliquet mi est nascetur interdum placerat
#> tellus ridiculus fusce.