Generate Lorem Ipsum Text

Generates lorem ipsum placeholder text for the requested number of sentences or paragraphs. You can control the number of sentences per paragraph and the average number of words per sentence, or simply enter the number of desired paragraphs for a completely random experience.

lorem::ipsum() uses sampling and the random number generator and makes no effort to shield the placeholder text generation from the main script, so please only use this package for temporary placeholder text.

Usage

ipsum(paragraphs = 1, sentences = NULL, avg_words_per_sentence = 10)

ipsum_words(n, collapse = TRUE)

ipsum_starts(n)

Arguments

paragraphs

Number of paragraphs of text to generate.

sentences

Number of sentences per paragraph. If NULL, then a random number of sentences per paragraph (approximately 3-8) will be chosen. Alternatively, sentences can be a vector of integers representing the number of sentences per paragraph.

avg_words_per_sentence

Number of expected words per sentence.

n

Number of words to generate

collapse

Should the words be collapsed into a single string, separated by spaces (default)? If FALSE, the chosen words are returned as a character vector.

Value

A character vector of lorem ispum placeholder text, where each element in the vector is a paragraph of text.

Functions

ipsum(): Generate paragraphs and sentences of lorem ipsum text.

ipsum_words(): Generate lorem ipsum words, without punctuation.

ipsum_starts(): Generate lorem ipsum starting words.

Options

You can influence, to a degree, the amount of punctuation that is included in the output using the lorem.punctuation_valence option. This global option should be a number between 0 and 1, or FALSE to disable punctuation altogether. When the value is closer to 1, more punctuation is included in the sentences. When the value is closer to 0, less punctuation will be inserted. The default value is 0.4.

Examples

# 1 paragraph of text
lorem::ipsum(1)
#> <lorem>
#> Amet arcu rutrum ridiculus posuere auctor placerat ullamcorper semper
#> ullamcorper. Placerat lacus eleifend maecenas condimentum blandit
#> magnis blandit cras. Aenean sodales nibh at suscipit ridiculus risus
#> per, fermentum curabitur nibh? Sed euismod sociosqu vivamus justo; duis
#> fames per.

# 2 paragraphs with 2 and 3 sentences each
lorem::ipsum(2, sentences = c(2, 3))
#> <lorem>
#> Adipiscing scelerisque ultricies pellentesque dignissim tincidunt
#> facilisis enim donec nullam quam. Taciti justo facilisi integer
#> sociosqu cubilia ut.
#> 
#> Sit dapibus, lacus, maecenas justo venenatis ut odio eu sagittis. Lacus
#> ligula vitae potenti dictum imperdiet litora luctus vulputate. Lacus
#> nulla natoque aliquet placerat platea conubia tempor ligula fusce
#> feugiat: facilisi, accumsan, eget tristique nam venenatis montes
#> platea.

# 2 paragraphs with short sentences
lorem::ipsum(2, avg_words_per_sentence = 4)
#> <lorem>
#> Ipsum tempor quisque rhoncus aenean consequat pharetra litora rhoncus
#> nibh nam!
#> 
#> Adipiscing odio magnis fringilla mus. Nostra aliquam quam, magna
#> sodales. Mi tellus pulvinar nunc senectus! Mattis commodo ac nullam
#> tincidunt auctor. Nullam nisi ultrices praesent aptent lectus. Na
#> lectus.

# 2 paragraphs with long sentences
lorem::ipsum(2, avg_words_per_sentence = 20)
#> <lorem>
#> Lorem venenatis facilisis massa massa – leo urna ultrices platea.
#> Venenatis penatibus velit accumsan conubia eros. Vitae turpis cursus
#> massa aenean netus non inceptos montes interdum volutpat pellentesque
#> dui faucibus! Tempus risus enim mus viverra conubia augue eleifend
#> torquent. Nostra justo luctus pharetra, litora iaculis bibendum mauris
#> cubilia conubia donec. Nascetur varius duis sollicitudin justo facilisi
#> fermentum litora libero conubia libero curabitur proin taciti phasellus
#> eget dictumst nec consequat maecenas blandit velit quis ultricies
#> scelerisque tempus at sociis, pharetra nunc mauris lobortis donec
#> phasellus, donec facilisis, lacus mollis parturient id eget nisi, leo,
#> class ac habitasse nisl quam orci ornare varius, massa diam mus primis
#> feugiat fringilla dis hendrerit montes habitant hendrerit felis nostra
#> commodo sodales scelerisque tincidunt litora elementum magna at sem
#> pharetra donec dignissim!
#> 
#> Lorem fames malesuada, fermentum ac torquent vehicula sollicitudin
#> rutrum iaculis curae nibh nunc. Odio malesuada eu torquent justo id
#> vel, mollis, congue ultricies massa turpis platea lectus? Litora
#> quisque fringilla proin montes arcu fusce sociis maecenas aptent?
#> Scelerisque magnis ultrices eleifend himenaeos faucibus parturient,
#> fermentum mauris ac eros dignissim. Ante ante rhoncus mollis lobortis
#> sociis porta suspendisse cras cras facilisis venenatis dictumst erat ut
#> etiam commodo mattis justo habitasse lectus dui nulla: natoque, conubia
#> ullamcorper – ornare – augue interdum diam, varius, sodales placerat,
#> nibh turpis lacinia litora et ac dis sed donec augue inceptos tincidunt
#> odio aptent nisi lacus netus curae scelerisque augue condimentum
#> sagittis cras.