Generate Lorem Ipsum Text

Generates lorem ipsum placeholder text for the requested number of sentences or paragraphs. You can control the number of sentences per paragraph and the average number of words per sentence, or simply enter the number of desired paragraphs for a completely random experience. Note that lorem::ipsum() uses the random number generator and sampling and makes no effort to shield the placeholder text generation from the main script, so please only use this package for temporary placeholder text.

Usage

ipsum(paragraphs = 1, sentences = NULL, avg_words_per_sentence = 10)

ipsum_words(n, collapse = TRUE)

ipsum_starts(n)

Arguments

paragraphs

Number of paragraphs of text to generate.

sentences

Number of sentences per paragraph. If NULL, then a random number of sentences per paragraph (approximately 3-8) will be chosen. Alternatively, sentences can be a vector of integers representing the number of sentences per paragraph.

avg_words_per_sentence

Number of expected words per sentence.

n

Number of words to generate

collapse

Should the words be collapsed into a single string, separated by spaces (default)? If FALSE, the chosen words are returned as a character vector.

Value

A character vector of lorem ispum placeholder text, where each element in the vector is a paragraph of text.

Functions

ipsum(): Generate paragraphs and sentences of lorem ipsum text.

ipsum_words(): Generate lorem ipsum words, without punctuation.

ipsum_starts(): Generate lorem ipsum starting words.

Options

You can influence, to a degree, the amount of punctuation that is included in the output using the lorem.punctuation_valence option. This global option should be a number between 0 and 1, or FALSE to disable punctuation altogether. When the value is closer to 1, more punctuation is included in the sentences. When the value is closer to 0, less punctuation will be inserted. The default value is 0.4.

Examples

# 1 paragraph of text
lorem::ipsum(1)
#> <lorem>
#> Ipsum tempus cum quam, iaculis mauris auctor mollis; condimentum augue
#> inceptos nisi? Euismod integer pharetra velit lacus dictumst nullam
#> ornare a pretium vitae. Phasellus aenean justo, dictum dictumst: sociis
#> mollis vulputate elementum ad taciti suscipit.

# 2 paragraphs with 2 and 3 sentences each
lorem::ipsum(2, sentences = c(2, 3))
#> <lorem>
#> Dolor vitae a nam dictumst porttitor, pretium, non bibendum. Ad
#> convallis et euismod condimentum non?
#> 
#> Sit inceptos nec vitae faucibus luctus magnis taciti. Dis nisl tempor,
#> cursus nascetur elementum at hac porttitor! Habitasse auctor conubia
#> elementum laoreet nascetur dapibus taciti, eros massa sapien sem luctus
#> odio fusce.

# 2 paragraphs with short sentences
lorem::ipsum(2, avg_words_per_sentence = 4)
#> <lorem>
#> Adipiscing nullam aliquam ridiculus. Magna tempor nostra diam. Nullam
#> vel potenti cum? In sagittis risus quisque cum. Accumsan sagittis
#> turpis curae cursus! Tempor nullam condimentum conubia scelerisque? Na
#> scelerisque.
#> 
#> Lorem facilisis per metus lacinia! Dapibus faucibus aliquet sed
#> phasellus. Nam elementum rutrum morbi turpis. Eleifend lobortis posuere
#> risus iaculis. Aliquet quis diam pretium mattis lacinia? Purus
#> facilisis curabitur viverra nulla varius? Parturient inceptos suscipit
#> ut elementum hac. Na hac?

# 2 paragraphs with long sentences
lorem::ipsum(2, avg_words_per_sentence = 20)
#> <lorem>
#> Ipsum malesuada praesent ornare sociosqu magna – convallis nisl
#> tincidunt ac id. Conubia venenatis fringilla semper urna interdum
#> ullamcorper fusce condimentum et. Dictumst fames ultrices blandit
#> vulputate suspendisse taciti nam fermentum, quis praesent. Ad id
#> blandit penatibus iaculis posuere vestibulum sollicitudin a justo
#> commodo nunc eros, porttitor, cursus proin a commodo elementum odio,
#> porta habitasse non porttitor placerat sagittis dictumst pulvinar est
#> massa est tempor per sociosqu pretium!
#> 
#> Sit accumsan ultricies curae mi, orci: convallis condimentum donec
#> sodales sem. Nibh tempor molestie ad class, nulla nisi malesuada? Justo
#> volutpat nisi dictum varius vestibulum ad sociosqu, purus a montes
#> gravida? Curae euismod est viverra mattis cubilia ullamcorper, mi
#> libero mus! Varius suscipit fames libero maecenas varius sem blandit
#> eros convallis mauris pharetra sed; porttitor fermentum vulputate odio
#> justo pretium pulvinar bibendum velit odio nulla suscipit nisl ultrices
#> fames, varius, aliquam nulla aliquam potenti habitant, diam nec
#> sollicitudin, vestibulum dictumst venenatis integer id venenatis
#> hendrerit, enim suscipit per eu condimentum non pellentesque consequat
#> ultrices fringilla potenti tristique commodo non fusce primis justo
#> odio inceptos?