Generate Lorem Ipsum Text

Generates lorem ipsum placeholder text for the requested number of sentences or paragraphs. You can control the number of sentences per paragraph and the average number of words per sentence, or simply enter the number of desired paragraphs for a completely random experience. Note that lorem::ipsum() uses the random number generator and sampling and makes no effort to shield the placeholder text generation from the main script, so please only use this package for temporary placeholder text.

Usage

ipsum(paragraphs = 1, sentences = NULL, avg_words_per_sentence = 10)

ipsum_words(n, collapse = TRUE)

ipsum_starts(n)

Arguments

paragraphs

Number of paragraphs of text to generate.

sentences

Number of sentences per paragraph. If NULL, then a random number of sentences per paragraph (approximately 3-8) will be chosen. Alternatively, sentences can be a vector of integers representing the number of sentences per paragraph.

avg_words_per_sentence

Number of expected words per sentence.

n

Number of words to generate

collapse

Should the words be collapsed into a single string, separated by spaces (default)? If FALSE, the chosen words are returned as a character vector.

Value

A character vector of lorem ispum placeholder text, where each element in the vector is a paragraph of text.

Functions

ipsum(): Generate paragraphs and sentences of lorem ipsum text.

ipsum_words(): Generate lorem ipsum words, without punctuation.

ipsum_starts(): Generate lorem ipsum starting words.

Options

You can influence, to a degree, the amount of punctuation that is included in the output using the lorem.punctuation_valence option. This global option should be a number between 0 and 1, or FALSE to disable punctuation altogether. When the value is closer to 1, more punctuation is included in the sentences. When the value is closer to 0, less punctuation will be inserted. The default value is 0.4.

Examples

# 1 paragraph of text
lorem::ipsum(1)
#> <lorem>
#> Sit pretium sociis; hendrerit mollis torquent lobortis; quam cursus
#> nostra lacinia natoque? Risus sapien cursus posuere, sodales mauris
#> class; nostra nibh convallis. Fringilla donec libero augue, curae
#> senectus malesuada tellus feugiat leo class sapien. Mauris curae nullam
#> dignissim gravida bibendum ut condimentum volutpat vitae. Curae augue
#> ante dignissim facilisis semper, rhoncus lacinia, faucibus penatibus
#> eget.

# 2 paragraphs with 2 and 3 sentences each
lorem::ipsum(2, sentences = c(2, 3))
#> <lorem>
#> Sit nisl mi ridiculus proin. Vivamus fermentum dis orci lacinia sociis
#> quis id vel venenatis penatibus eu; etiam neque!
#> 
#> Elit potenti orci ridiculus aptent molestie bibendum, turpis
#> suspendisse: dictum lacus senectus curabitur pretium faucibus. Primis
#> feugiat tempor tincidunt rutrum sollicitudin, viverra sodales fermentum
#> vestibulum. Lacinia sodales quis nunc inceptos; primis habitant
#> condimentum fringilla.

# 2 paragraphs with short sentences
lorem::ipsum(2, avg_words_per_sentence = 4)
#> <lorem>
#> Ipsum lacinia etiam aptent suspendisse. Parturient sociosqu torquent
#> duis class. Dui fames convallis accumsan magnis tristique. Na
#> tristique.
#> 
#> Ipsum consequat nullam taciti. Erat id consequat imperdiet. Sapien
#> ullamcorper dui placerat eleifend? Velit urna sagittis conubia
#> torquent. Cras neque faucibus tempus tincidunt! Est inceptos parturient
#> condimentum fermentum! Na fermentum!

# 2 paragraphs with long sentences
lorem::ipsum(2, avg_words_per_sentence = 20)
#> <lorem>
#> Amet vestibulum nostra, vitae nam natoque bibendum consequat nisl
#> sodales, taciti consequat purus aliquet. Potenti penatibus laoreet
#> facilisis venenatis porta nisi curae: fermentum eros cursus gravida.
#> Posuere molestie tempor nunc conubia vel libero ornare tempus. In
#> blandit natoque, duis nunc cubilia est pretium. Nibh pellentesque
#> accumsan lacinia dictumst malesuada! Duis ridiculus viverra ante mus
#> in, venenatis gravida risus. Lacus scelerisque neque et tempus eget
#> suscipit facilisis diam suscipit sem rhoncus erat euismod diam hac
#> molestie sem netus fames, interdum non himenaeos class quisque nulla
#> consequat interdum diam massa nostra cras fringilla: leo, non donec,
#> risus fermentum malesuada, montes odio luctus eros massa, quis sodales
#> mauris vulputate nisl inceptos est iaculis vitae arcu laoreet:
#> himenaeos nam maecenas elementum pretium varius sapien tellus nam
#> euismod ad erat a volutpat augue sociis aenean penatibus dignissim
#> lacinia mus elementum rutrum.
#> 
#> Lorem penatibus facilisis – viverra maecenas neque himenaeos id class:
#> felis justo taciti? Fusce luctus mollis, fusce metus sem sociis aptent
#> risus, facilisis, torquent leo. Integer id fusce pharetra, lobortis
#> risus morbi. A tortor lectus nisl, dis conubia. Lectus dapibus nascetur
#> id non pretium leo porttitor eu non urna curabitur sodales porttitor
#> ullamcorper vestibulum nisl himenaeos – bibendum cubilia cursus tempor
#> bibendum mi sodales vel lacus potenti aliquam; gravida vitae ligula,
#> hac lobortis pulvinar, cursus luctus volutpat dui mi ante pellentesque
#> habitant accumsan curabitur felis ut, scelerisque morbi nam condimentum
#> eu urna porta et quam porta auctor iaculis commodo blandit at netus
#> dictumst quam a rhoncus interdum duis sociosqu?