Generate Lorem Ipsum Text

Generates lorem ipsum placeholder text for the requested number of sentences or paragraphs. You can control the number of sentences per paragraph and the average number of words per sentence, or simply enter the number of desired paragraphs for a completely random experience.

lorem::ipsum() uses sampling and the random number generator and makes no effort to shield the placeholder text generation from the main script, so please only use this package for temporary placeholder text.

Usage

ipsum(paragraphs = 1, sentences = NULL, avg_words_per_sentence = 10)

ipsum_words(n, collapse = TRUE)

ipsum_starts(n)

Arguments

paragraphs

Number of paragraphs of text to generate.

sentences

Number of sentences per paragraph. If NULL, then a random number of sentences per paragraph (approximately 3-8) will be chosen. Alternatively, sentences can be a vector of integers representing the number of sentences per paragraph.

avg_words_per_sentence

Number of expected words per sentence.

n

Number of words to generate

collapse

Should the words be collapsed into a single string, separated by spaces (default)? If FALSE, the chosen words are returned as a character vector.

Value

A character vector of lorem ispum placeholder text, where each element in the vector is a paragraph of text.

Functions

ipsum(): Generate paragraphs and sentences of lorem ipsum text.

ipsum_words(): Generate lorem ipsum words, without punctuation.

ipsum_starts(): Generate lorem ipsum starting words.

Options

You can influence, to a degree, the amount of punctuation that is included in the output using the lorem.punctuation_valence option. This global option should be a number between 0 and 1, or FALSE to disable punctuation altogether. When the value is closer to 1, more punctuation is included in the sentences. When the value is closer to 0, less punctuation will be inserted. The default value is 0.4.

Examples

# 1 paragraph of text
lorem::ipsum(1)
#> <lorem>
#> Consectetur pharetra sagittis urna magnis dapibus id semper nullam.
#> Consequat magna urna inceptos aptent, cubilia purus mattis – platea
#> nascetur penatibus; justo etiam!

# 2 paragraphs with 2 and 3 sentences each
lorem::ipsum(2, sentences = c(2, 3))
#> <lorem>
#> Ipsum sociis interdum eros condimentum mauris vestibulum tortor
#> pellentesque nulla nisi cras ullamcorper. Curae ante mattis tempor
#> lacinia curabitur integer facilisi suscipit nibh dictum.
#> 
#> Adipiscing massa mus sagittis potenti suscipit ornare himenaeos
#> pellentesque nec fermentum. Purus enim curabitur gravida magnis
#> imperdiet. Nostra placerat torquent vitae, egestas vulputate tincidunt
#> hac, sem sapien scelerisque suscipit rutrum netus vitae justo!

# 2 paragraphs with short sentences
lorem::ipsum(2, avg_words_per_sentence = 4)
#> <lorem>
#> Elit condimentum eros feugiat imperdiet. Sodales tincidunt vestibulum
#> vulputate conubia curae. Vehicula nec primis vehicula, pulvinar netus.
#> Na netus.
#> 
#> Ipsum nam duis lacus. Duis diam praesent at nibh. Faucibus sagittis
#> sodales, turpis aptent? Vulputate faucibus per a odio. Molestie risus
#> eleifend ornare congue. Na congue?

# 2 paragraphs with long sentences
lorem::ipsum(2, avg_words_per_sentence = 20)
#> <lorem>
#> Amet nisl habitasse dis urna magna lacinia elementum commodo. Elementum
#> ullamcorper per, bibendum quisque tempus. Porttitor penatibus porta
#> facilisi potenti, urna aptent. Libero diam feugiat – convallis
#> pellentesque non. Blandit pharetra aptent risus massa eu odio fermentum
#> tortor nullam fringilla commodo dui aenean arcu hac ante id odio proin
#> aptent sociis aenean, faucibus turpis orci tristique a mattis aliquam
#> interdum, ultricies, quis, varius, et urna, conubia arcu senectus,
#> iaculis malesuada nam nascetur libero faucibus molestie hac curabitur
#> leo bibendum curabitur feugiat fringilla gravida lectus quis ut id nam
#> congue primis nascetur aliquam morbi sem nullam dictum praesent nam hac
#> quam dui lobortis nostra enim sem eget porta placerat.
#> 
#> Consectetur tincidunt bibendum mi; velit himenaeos, pulvinar per dui
#> semper eros curabitur netus. Porta faucibus lacinia platea cursus
#> lobortis; odio curae. Hac urna eros vestibulum vulputate posuere mattis
#> in parturient ultricies, nulla ac. Eleifend ultrices elementum
#> condimentum, neque blandit dictum sociis rhoncus. Turpis ultricies
#> sodales netus cum lacus id platea nascetur elementum sed: ridiculus
#> auctor imperdiet natoque. Bibendum mattis neque vulputate netus nisl
#> vehicula turpis lacus feugiat egestas orci pellentesque morbi convallis
#> elementum accumsan sed integer taciti turpis laoreet porta hac netus
#> leo faucibus felis auctor: fermentum feugiat: ad hac odio sociis cras
#> tempor fusce nascetur quam habitasse himenaeos tortor orci rhoncus
#> lobortis pharetra magna ultrices in torquent tempus etiam.